+++ src/plone/api/__init__.py
@@ -1,1 +1,1 @@
-from plone.api import content, user, group, portal  # noQA
+from plone.api import content, user, group, portal, env  # noQA